+++ internal/provider/fwprovider/provider.go
@@ -1,4 +1,4 @@
-package fwserver
+package fwprovider
 
 import (
 	"context"
@@ -11,13 +11,13 @@ import (
 )
 
 func New() tfsdk.Provider {
-	return &providerServer{}
+	return &provider{}
 }
 
-type providerServer struct{}
+type provider struct{}
 
 // GetSchema returns the schema for this provider's configuration.
-func (p *providerServer) GetSchema(ctx context.Context) (tfsdk.Schema, diag.Diagnostics) {
+func (p *provider) GetSchema(ctx context.Context) (tfsdk.Schema, diag.Diagnostics) {
 	var diags diag.Diagnostics
 
 	// This schema must match exactly the Terraform Protocol v5 (Terraform Plugin SDK v2) provider's schema.
@@ -285,12 +285,12 @@ func (p *providerServer) GetSchema(ctx context.Context) (tfsdk.Schema, diag.Diag
 // Configure is called at the beginning of the provider lifecycle, when
 // Terraform sends to the provider the values the user specified in the
 // provider configuration block.
-func (p *providerServer) Configure(ctx context.Context, request tfsdk.ConfigureProviderRequest, response *tfsdk.ConfigureProviderResponse) {
+func (p *provider) Configure(ctx context.Context, request tfsdk.ConfigureProviderRequest, response *tfsdk.ConfigureProviderResponse) {
 }
 
 // GetResources returns a mapping of resource names to type
 // implementations.
-func (p *providerServer) GetResources(ctx context.Context) (map[string]tfsdk.ResourceType, diag.Diagnostics) {
+func (p *provider) GetResources(ctx context.Context) (map[string]tfsdk.ResourceType, diag.Diagnostics) {
 	var diags diag.Diagnostics
 	resources := make(map[string]tfsdk.ResourceType)
 
@@ -299,7 +299,7 @@ func (p *providerServer) GetResources(ctx context.Context) (map[string]tfsdk.Res
 
 // GetDataSources returns a mapping of data source name to types
 // implementations.
-func (p *providerServer) GetDataSources(ctx context.Context) (map[string]tfsdk.DataSourceType, diag.Diagnostics) {
+func (p *provider) GetDataSources(ctx context.Context) (map[string]tfsdk.DataSourceType, diag.Diagnostics) {
 	var diags diag.Diagnostics
 	dataSources := make(map[string]tfsdk.DataSourceType)
